requirements.txt
flask
google-cloud-bigquery
pandas
db-dtypes
openai
python-dotenv
matplotlib
plotly